6-size.c updated

+++ 0x00-hello_world/6-size.c
@@ -6,9 +6,9 @@
  */
 int main(void)
 {
-	printf("The size of int: %d bytes\n", (int)sizeof(int));
-	printf("The size of char: %d bytes\n", (int)sizeof(char));
-	printf("The size of float: %d byts\n", (int)sizeof(float));
-	printf("The size of long: %d bytes\n", (int)sizeof(long));
+	printf("The size of int: %d bytes\n", sizeof(int));
+	printf("The size of char: %d bytes\n", sizeof(char));
+	printf("The size of float: %d byts\n", sizeof(float));
+	printf("The size of long: %d bytes\n", sizeof(long));
 	return (0);
 }
